feat(wiki): safe headless wiki-authoring drain — async, throttled, anti-fabrication
Make the headless wiki-authoring drain safe to enable: rewrite the
claude -p invocation as async (asyncio.create_subprocess_exec +
wait_for), bound it with a per-cycle budget (concurrency / wall-clock /
USD caps via CORTEX_HEADLESS_* env knobs), and gate authoring behind a
groundable filter so non-derivable scopes (prd/decisions/changelog/
roadmap/accessibility/localization) are never fabricated.

Reviewer fixes (REQUEST CHANGES → resolved):
- BLOCK-1: _claude_invoke now kills the subprocess in a finally block
  when returncode is None — CancelledError (a BaseException) escaped the
  except clauses and leaked a zombie; the timeout path shares the kill.
- BLOCK-3: _write_anchor_page mkdir moved inside the try/except OSError
  so a PermissionError cannot propagate through gather(return_exceptions
  =False) and abort the whole anchor phase.
- BLOCK-2: split the 1766-line module into 6 focused modules (all ≤500),
  keeping headless_authoring as the public import surface; cycle code
  resolves monkeypatchable names from the root module at call time.
- MAJOR-1: _collect_anchor_candidates wraps _build_registry().repos in
  its try/except.
- MINOR-1/2, NIT-1: document wall_clock_ms ≡ duration_ms; drains_attempted
  excludes budget-skips; deterministic wall-clock test (monkeypatch
  CycleBudget.exhausted instead of real-sleep timing).

Security controls (B-1) intact: --bare, --tools Read,Glob,Grep,
ANTHROPIC_API_KEY fail-closed. 51 tests pass.
